Parralelize SVMTool learning and analysis.

+++ lima_pelf/evalPosTagging/launch_eval.sh
@@ -4,7 +4,8 @@ set -o nounset
 set -o errexit
 set -o pipefail
 
-export PATH=/home/sid-ahmed/Lima/lima/lima_pelf/evalPosTagging/SVM/SVMTool-1.3.1/bin:$PATH
+command -v SVMTlearn >/dev/null 2>&1 || { echo >&2 "I require SVMTlearn but it's not in the path  Aborting."; exit 1; }
+command -v svm_learn >/dev/null 2>&1 || { echo >&2 "I require svm_learn but it's not in the path  Aborting."; exit 1; }
 
 if (($#==0)); then 
     echo usage: launch_eval.sh lang1 lang2 ...
@@ -18,15 +19,18 @@ nbParts=10
 EVAL_PATH=`dirname $0`
 
 # svm_light location:
-svm_light="/usr/local/bin";
+svm_light=$(dirname $(command -v svm_learn ))
 # svm_learn location (perl version):
 svm_learn="SVMTlearn"
 # svm_learn parameters:
 # sed -i "s/use lib \".*\/lib/\";/use lib \"$EVAL_PATH\/SVM\/SVMTool-1.3.1\/lib\/\";/g" SVM/SVMTool-1.3.1/bin/SVMTlearn
+
+
 # make sure training-sets exists
-rm -Rf training-sets
+# rm -Rf training-sets
 mkdir -p training-sets
 
+notrain=false
 for lang in $*; do
     addOption=""
     case $lang in
@@ -37,15 +41,21 @@ for lang in $*; do
             conf=config-minimale-eng.SVMT;;
         fre) corpus=$LINGUISTIC_DATA_ROOT/disambiguisationMatrices/fre/corpus/corpus_fre.txt; conf=config-minimale-fre.SVMT ;;
         por) addOption="-s PU+FORTE -n $nbParts"; corpus=$LINGUISTIC_DATA_ROOT/disambiguisationMatrices/por/corpus/macmorpho.conll.txt; conf=config-minimale-por.SVMT ;;
+        notrain) notrain=true ;;
     esac
     method=$(readMethod $lang)
     echo "treating $lang.$method... (see ${lang}.${method}.log)"
 
-    rm -Rf results.$lang.$method
 
     confFile=$(findFileInPaths $LIMA_CONF $conf ":")
-
-    $EVAL_PATH/tfcv.py -c -t -l $lang $addOption $corpus $confFile $svm_light $svm_learn
+    
+    if [ $notrain = true ]
+    then
+        $EVAL_PATH/tfcv.py -l $lang $addOption $corpus $confFile $svm_light $svm_learn
+    else
+        rm -Rf results.$lang.$method
+        $EVAL_PATH/tfcv.py -c -t -l $lang $addOption $corpus $confFile $svm_light $svm_learn
+    fi
     if  [[ $? -ne 0 ]]; then
       echo "tfcv error, exiting"
       exit 1
